src/user_service/requirement.txt
Django==5.2
djangorestframework>=3.15
psycopg2-binary>=2.9.9
gunicorn>=22.0.0
django-cors-headers>=4.4.0
python-dotenv>=1.0.1
django-prometheus>=2.3.1
djangorestframework-simplejwt==5.5.1